python/projects/py-pia/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pia"
version = "0.1.0"
description = "Product Intelligence Agent - CLI tool for tracking and analyzing database product releases"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12",
    "pydantic>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "rich>=13.0",
    "python-dateutil>=2.9",
    "markdownify>=0.13",
]

[project.scripts]
pia = "pia.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
